Renaming version control to git
