+++ projects/utils/src/api/git/profile.ts
@@ -1,0 +1,8 @@
+import { IGitSystem } from './api'
+import { LibraryProfile } from '../../types'
+
+export const gitProfile: LibraryProfile<IGitSystem> = {
+  name: 'git',
+  methods: ['clone', 'checkout', 'init', 'add', 'commit', 'fetch', 'pull', 'push', 'reset', 'status', 'remote', 'log']
+  //events: ['currentFileChanged']
+}
